New asteroids!
We might have to change a little bit with the camera, but other than that these asteroids are in and are ready for deployment! How will we go about giving them movement? Just put them on splines like I did with the camera in the mainmenu map for NWV, or will we have a more randomized approach to it?