HANZ SOUNDS
He makes noises when he speaks.
Damage for if he is hit by a meteor.
Failure for if the ship explodes or if the sun explodes.
Victory or if the mission is a success.
ThatWasInteresting: No idea. It was candid and accidental.
HanzSolarHaulin: Should be coupled with a video demonstration.